refactor: RepositoryTest 를 통합한다.
